backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
